requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async HTTP client
aiohttp>=3.9.2
aiofiles==23.2.1

# Environment variables
python-dotenv==1.0.0

# Data handling
orjson==3.9.12

# CORS
fastapi-cors==0.0.6

# SSE support
sse-starlette==1.8.2